fix(oss-accel): opcode 类型不匹配致 make build 静默失败(cfs-server 从未真正重建)
真相大白:此前 rc12/rc13/rc14 全部构建于陈旧 cfs-server 二进制——opcode
声明为裸 int(:=推断),mp.submit 要 uint32,make build 对 metanode
包报编译错误 failed,但 make image 的 Dockerfile COPY build/bin/ 不
检查上一步是否成功,盲目打包了更早一次构建残留的旧二进制;我用
tail -N 截断 make build 输出,恰好把这条 failed 错误滤没了。

修:opcode := uint32(opFSMUpdateExtentKeyAfterMigration)。

+++ metanode/partition_op_inode.go
@@ -1207,7 +1207,7 @@ func (mp *metaPartition) UpdateExtentKeyAfterMigration(req *proto.UpdateExtentKe
 		return
 	}
 
-	opcode := opFSMUpdateExtentKeyAfterMigration
+	opcode := uint32(opFSMUpdateExtentKeyAfterMigration)
 	if req.ColdBackendExternal && req.StorageClass == proto.StorageClass_BlobStore {
 		opcode = opFSMUpdateExtentKeyAfterMigrationColdExternal
 	}
